+++ packages/wallet/index.ts
@@ -1,4 +1,4 @@
-import '@rerock/polyfills'
+import '@rerock/appkit-polyfills'
 export { W3mFrame } from './src/W3mFrame.js'
 export { W3mFrameHelpers } from './src/W3mFrameHelpers.js'
 export { W3mFrameProvider } from './src/W3mFrameProvider.js'