+++ packages/solana/CHANGELOG.md
@@ -13,6 +13,6 @@
   - @rerock/appkit@0.0.3
   - @rerock/appkit-common@0.0.3
   - @rerock/appkit-core@0.0.3
-  - @rerock/polyfills@0.0.3
+  - @rerock/appkit-polyfills@0.0.3
   - @rerock/scaffold-utils@0.0.3
   - @rerock/wallet@0.0.3