+++ packages/wagmi/CHANGELOG.md
@@ -12,7 +12,7 @@
   - @rerock/appkit-adapter-wagmi@0.0.3
   - @rerock/appkit@0.0.3
   - @rerock/appkit-common@0.0.3
-  - @rerock/polyfills@0.0.3
+  - @rerock/appkit-polyfills@0.0.3
   - @rerock/scaffold-utils@0.0.3
   - @rerock/siwe@0.0.3
   - @rerock/wallet@0.0.3