+++ packages/wallet/CHANGELOG.md
@@ -10,4 +10,4 @@
 
 - Updated dependencies [[`91d0296`](https://github.com/WalletConnect/shadow-appkit/commit/91d02963cbe3c2d06b74801b519ce23dd30ff797), [`51eff9f`](https://github.com/WalletConnect/shadow-appkit/commit/51eff9f82c296b0ba2b5ab33af92a1fa54a77f7a)]:
   - @rerock/appkit-common@0.0.3
-  - @rerock/polyfills@0.0.3
+  - @rerock/appkit-polyfills@0.0.3